!XB (Sandbox) (CE-20735) False positive "cannot delete asset file" warning for materials created within the root of GameSDK. False positive "material with duplicate name" when creating a material with the name of the previously deleted material.
Copied from Perforce
 Change: 1996871

+++ Code/Sandbox/Plugins/EditorCommon/AssetSystem/AssetManager.cpp
@@ -387,7 +387,7 @@ void CAssetManager::MergeAssets(std::vector<CAsset*> assets)
 				editable.SetDetails(assets[i]->GetDetails());
 
 				// Force to update the thumbnail.
-				if (!pExistingAsset->IsImmutable())
+				if (pExistingAsset->IsThumbnailLoaded() && !pExistingAsset->IsImmutable())
 				{
 					pPak->RemoveFile(pExistingAsset->GetThumbnailPath());
 				}
